Fix(Renderer): Improve stability and collision detection
This commit introduces several fixes and improvements to the renderer:

-   Refined collision detection logic in  for improved accuracy,
    especially for circle-rectangle and rectangle-rectangle intersections.
-   Changed Collider constructor parameters to use floats for better precision.
-   Added a  method to the  class and exposed it
    through the  for better shape management.
-   Adjusted destructor logic in  and  to
    address potential memory issues or double-frees related to colliders
    and shapes.
-   Removed  and its inclusion, suggesting a change in
    profiling strategy or its removal.

+++ main/espFeatures/renderer/include/Collider.hpp
@@ -43,7 +43,7 @@ class Collider : public Collidable {
     float x;
     float y;
 
-    Collider(int x, int y) : x(x), y(y) {}
+    Collider(float x, float y) : x(x), y(y) {}
     virtual ~Collider() = default;
 
     virtual bool intersects(const Collider *other) const = 0;
